Move WorkspaceDeletedBanner
