+++ site/src/pages/WorkspacePage/Workspace.tsx
@@ -16,7 +16,7 @@ import { Margins } from "components/Margins/Margins"
 import { Resources } from "components/Resources/Resources"
 import { Stack } from "components/Stack/Stack"
 import { WorkspaceActions } from "pages/WorkspacePage/WorkspaceActions/WorkspaceActions"
-import { WorkspaceDeletedBanner } from "components/WorkspaceDeletedBanner/WorkspaceDeletedBanner"
+import { WorkspaceDeletedBanner } from "./WorkspaceDeletedBanner"
 import { WorkspaceStats } from "components/WorkspaceStats/WorkspaceStats"
 import {
   FullWidthPageHeader,
