+++ site/src/pages/WorkspacePage/WorkspaceDeletedBanner.stories.tsx
