+++ site/src/pages/WorkspacePage/WorkspaceDeletedBanner.tsx
